refactor: isolate history requests and shared cache primitives
